Enhancements to calendar frontend and functionality
I got rid of the tasks section and replaced the goals section as the "tasks" section as requested by the client. I also created the functionality where when the dragging the draggable event, when it hovers over the date, then it will light up so that the user knows exactly which date they will put the event in. Also, there is a marker for which date has a task, which improves usability.