Developed the stats change functionality
Developed the functionality where when the user clicks a button to complete the task, a new document called "stats" will be created under the user and will have the 5 types of tasks as fields that will be initially set to an integer 0. Depending on the type of task that is submitted, the type field on the stats doc will increase by 1, and these user stats are linked to a radar chart that will show up in the dashboard and the stats page.