Addition of Password reset and pet features
Added a password reset feature that uses firebase authentication functions. Created a pet.html page and implemented a system using javascript, where the use's stats data will be fetched and compared to each other, and the lowest will be sent on a modal notification popup that will remind the user to increase their stats in the task type in which they are lacking in.