Addition of Leaderboard + Overdue Tasks + Modification to low stats notification
Added the leaderboard in the dashboard page that shows the user first name, their level, and XP, and their rank in descending order. Also added the overdue tasks feature to show the user a modal that shows their overdue tasks, and also modified the low pet stats notification by converting the stats into an array and sorting the array to determine the lowest value, instead of having a constant integer for the threshold.